+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on http://keepachangelog.com/en/1.0.0/
 and this project adheres to http://semver.org/spec/v2.0.0.html.
 
+## [unreleased]
+
+- cmd/sense executes ./.onchange.sh if found
+- Add options to cmd/sense
+
 ## [0.2.0] 2022-07-17
 
 - Add constructor NewSensor and hide fields Root and Visit